Regenerate workflows
* Regenerated Minimal DPL workflow

* Regenerated QC DAQ Task workflow (general FLP QC)

* Regenerated PHOS compressor workflow

* Regenerated HMPID Raw QC Task workflow

* Regenerated TOF compressor workflows for FLP 178 and 179

* Use workflow inclusion, delete old stuff

* MFT decoder workflow

* MFT Raw QC Task workflow

* MFT digits QC Task + decoder

* Disable FairMQ shmmonitor in STFB/S

As Alexey said:
"But on a more general note, I don't think it makes much sense that TFBuilder autolaunches monitor and DPL doesn't. Should either they both do or they don't. If the old issue with the zombies is still relevant, we need to solve it in some other way. Actually an optimization in the upcoming tag should reduce the number of zombies (if there are any) to at most one per session."

* MID raw parser workflow