Align hook calls to follow the changes in OCTRL-902
OCTRL-902 allows to place hooks before or after SOSOR, EOSOR, SOEOR and EOEOR timestamps are created.
This PR moves the hook triggers so that the order of plugin execution is preserved and the information the plugins need is still available.
Calls which are critical for a successful SOR and EOR are placed inside the SOSOR - EOEOR timestamp boundaries.
Similarly, some other calls at after_<event>+0 are moved to after_<event>-1 to contribute to the transition, even if technically it does not change the perceived behaviour at the moment.