Delete qc-server-workflow (#684)
There is no legitimate use case for it anymore. Originally it was supposed to run long running QC workflows, which should now use Nomad instead.

This workflow template was never meant for testing QC workflows which are supposed to run in readout-dataflow. One should use the staging setup for trying out new configurations.