[OCTRL-1064] GO_ERROR calls should never be critical
Having to handle a failed GO_ERROR transition is something we should avoid. An interrupted GO_ERROR means that we might not perform the complete procedure and miss e.g. calling some services/subsystems.

With this change, the only known way of failing GO_ERROR is to try it in an inappropriate state, which can be ERROR (no point in transitioning from ERROR to ERROR) or DONE (too late, nothing to do).